Changes to make the repository upm package friendly
