+++ Assets/Controller2D/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7b67e81187be01c4d9da4f9654cf2b53
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
