+++ Assets/Controller2D/Sprites/New Palette.prefab
