+++ Assets/Controller2D/net.fiveotwo.characterController.asmdef
@@ -1,0 +1,13 @@
+{
+    "name": "net.fiveotwo.characterController",
+    "references": [],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
