+++ Assets/Controller2D/package.json
@@ -1,0 +1,20 @@
+{
+    "name": "net.fiveotwo.characterController",
+    "displayName": "Character Controller 2D",
+    "version": "1.0.0",
+    "unity": "2019.1",
+    "description": "A simple Character Controller for 2D games that require tight movement. It requires a BoxCollider2D.",
+    "keywords": [
+        "character",
+        "controller",
+        "2d",
+        "physics-less"
+    ],
+    "category": "2D",
+    "dependencies": {},
+    "license": "MIT",
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/502Studios/Controller2D"
+    }
+}
