+++ Assets/Controller2D/package.json.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 3d0f17b9427f0c9458cd4113831e6e7e
-folderAsset: yes
-DefaultImporter:
+guid: 6a11bb8823a20ae488e98c4538bdb41b
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
