--- Assets/Controller2D/raia.characterController.asmdef
@@ -1,3 +1,0 @@
-﻿{
-	"name": "raia.Controller"
-}
